Judge course-role badges from the counts, not the stored rate
Verifying the freshly loaded production table turned up four racers whose
qualification depends on where the comparison happens. Postgres stores the
rate as real, so 42/60, 21/30 and 26/40 (twice) land on 0.6999999... and
0.6499999... Counting the table in SQL reports 110 escape / 69 wall;
counting the same rows in Python reports 112 / 71.

Scope, measured rather than assumed: the app compares in Python, and
psycopg hands back 0.7, so today's badges are identical either way -- 40
escape races and 15 wall races under both rules, and no displayed rate
changes (0 of 1156). I expected psycopg's auto-prepare to flip the value
to binary and break this; probing past the threshold it did not. So this
is a consistency trap for whoever next counts these rows in SQL -- an
analytics query, a sync script, the verification I just ran -- rather
than a live badge bug.

Derive the rate from course*_starts and course*_wins, which are stored as
integers, so the verdict cannot depend on the column's precision. Widen
the rate columns to DOUBLE PRECISION so the SQL-side answer agrees too.

+++ scripts/build_racer_course_role_stats.py
@@ -42,10 +42,16 @@ def ensure_schema(conn) -> None:
           window_days           INTEGER NOT NULL,
           course1_starts        INTEGER NOT NULL,
           course1_wins          INTEGER NOT NULL,
-          course1_win_rate      REAL,
+          -- DOUBLE PRECISION にするのは REAL が Postgres では 4 バイトになり、
+          -- 42/60 や 26/40 の「ちょうど 0.70 / 0.65」を 0.6999999... に丸めて
+          -- しまうため。閾値がその値ちょうどなので、境界の選手が SQL 側の
+          -- 比較で静かに落ちる (2026-09-04)。判定自体は整数の分母分子から
+          -- 出すので率は表示用だが、後から SQL で数えた人が別の答えを得る
+          -- 状態を残さない。SQLite では REAL 相当 (8 バイト) で従来どおり。
+          course1_win_rate      DOUBLE PRECISION,
           course2_starts        INTEGER NOT NULL,
           course2_nigashi_count INTEGER NOT NULL,
-          course2_nigashi_rate  REAL,
+          course2_nigashi_rate  DOUBLE PRECISION,
           updated_at            TEXT NOT NULL,
           PRIMARY KEY (snapshot_date, racer_number)
         );
